# [PERF] Replace Sprintf in PromQL to string conversion