Goes faster due to reduced memory allocation.

This PR is organized into 7 commits: the first two add a test case and some benchmarks.
The rest make printing faster, in increasing levels of complexity.

I'm not convinced we need maximum complexity; most of the benefit is achieved with the first two changes.

1. Replace some simple Sprintf with string concat. This goes faster because there is no runtime format parsing.

```
                                                            │  before.txt  │             after1.txt              │
                                                            │    sec/op    │    sec/op     vs base               │
ExprString/sum_by(code)_(task:errors:rate10s{job="s...-8      455.9n ± 16%   441.6n ±  0%   -3.13% (p=0.002 n=6)
ExprString/max(_100_*_(1_-_avg_by(instance)_(irate(...-8      2.002µ ±  5%   1.315µ ±  8%  -34.30% (p=0.002 n=6)
ExprString/http_requests_total{job="api-server",_gr...-8      1.737µ ±  6%   1.154µ ±  1%  -33.59% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...-8      3.536µ ±  6%   2.614µ ±  4%  -26.06% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...#01-8   3.520µ ±  2%   2.584µ ± 10%  -26.58% (p=0.002 n=6)
ExprString/label_replace(testmetric,_"dst",_"destin...-8      927.5n ±  1%   675.9n ±  1%  -27.13% (p=0.002 n=6)
geomean                                                       1.623µ         1.205µ        -25.78%
```

2. Replace some Sprintf with bytes.Buffer 

```
                                                            │  before.txt   │             after2.txt              │
                                                            │    sec/op     │    sec/op     vs base               │
ExprString/sum_by(code)_(task:errors:rate10s{job="s...-8       455.9n ± 16%   172.8n ±  1%  -62.10% (p=0.002 n=6)
ExprString/max(_100_*_(1_-_avg_by(instance)_(irate(...-8      2001.5n ±  5%   888.1n ±  1%  -55.63% (p=0.002 n=6)
ExprString/http_requests_total{job="api-server",_gr...-8      1737.0n ±  6%   920.8n ± 13%  -46.99% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...-8       3.536µ ±  6%   1.891µ ±  1%  -46.53% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...#01-8    3.520µ ±  2%   1.875µ ±  2%  -46.73% (p=0.002 n=6)
ExprString/label_replace(testmetric,_"dst",_"destin...-8       927.5n ±  1%   471.6n ±  1%  -49.15% (p=0.002 n=6)
geomean                                                        1.623µ         786.2n        -51.57%
```

3. Replace every Sprintf with bytes.Buffer - eliminate almost all string concatenation.

```
                                                            │  before.txt   │             after3.txt             │
                                                            │    sec/op     │   sec/op     vs base               │
ExprString/sum_by(code)_(task:errors:rate10s{job="s...-8       455.9n ± 16%   173.0n ± 2%  -62.04% (p=0.002 n=6)
ExprString/max(_100_*_(1_-_avg_by(instance)_(irate(...-8      2001.5n ±  5%   552.0n ± 3%  -72.42% (p=0.002 n=6)
ExprString/http_requests_total{job="api-server",_gr...-8      1737.0n ±  6%   640.6n ± 1%  -63.12% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...-8       3.536µ ±  6%   1.421µ ± 1%  -59.82% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...#01-8    3.520µ ±  2%   1.406µ ± 1%  -60.05% (p=0.002 n=6)
ExprString/label_replace(testmetric,_"dst",_"destin...-8       927.5n ±  1%   342.8n ± 3%  -63.05% (p=0.002 n=6)
geomean                                                        1.623µ         589.3n       -63.69%
```


4. The last optimisation is to print `VectorSelector` faster when matchers are sorted, which in turn requires a change to the parser to put the `__name__` matcher first, and to fix up all the tests that check for this.

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/promql/parser
cpu: Apple M2
                                                            │  before.txt   │              after.txt              │
                                                            │    sec/op     │    sec/op     vs base               │
ExprString/sum_by(code)_(task:errors:rate10s{job="s...-8       455.9n ± 16%   129.0n ± 19%  -71.70% (p=0.002 n=6)
ExprString/max(_100_*_(1_-_avg_by(instance)_(irate(...-8      2001.5n ±  5%   492.1n ±  9%  -75.41% (p=0.002 n=6)
ExprString/http_requests_total{job="api-server",_gr...-8      1737.0n ±  6%   604.6n ±  1%  -65.19% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...-8       3.536µ ±  6%   1.482µ ±  4%  -58.10% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...#01-8    3.520µ ±  2%   1.184µ ±  5%  -66.36% (p=0.002 n=6)
ExprString/label_replace(testmetric,_"dst",_"destin...-8       927.5n ±  1%   352.2n ±  1%  -62.03% (p=0.002 n=6)
geomean                                                        1.623µ         536.0n        -66.98%

                                                            │  before.txt  │              after.txt              │
                                                            │     B/op     │     B/op      vs base               │
ExprString/sum_by(code)_(task:errors:rate10s{job="s...-8        240.0 ± 0%     224.0 ± 0%   -6.67% (p=0.002 n=6)
ExprString/max(_100_*_(1_-_avg_by(instance)_(irate(...-8       1456.0 ± 0%     306.0 ± 0%  -78.98% (p=0.002 n=6)
ExprString/http_requests_total{job="api-server",_gr...-8        856.0 ± 0%     371.0 ± 0%  -56.66% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...-8      3.432Ki ± 0%   1.608Ki ± 0%  -53.13% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...#01-8   3.385Ki ± 0%   1.335Ki ± 0%  -60.56% (p=0.002 n=6)
ExprString/label_replace(testmetric,_"dst",_"destin...-8        720.0 ± 0%     352.0 ± 0%  -51.11% (p=0.002 n=6)
geomean                                                       1.147Ki          521.7       -55.58%

                                                            │ before.txt  │             after.txt             │
                                                            │  allocs/op  │ allocs/op   vs base               │
ExprString/sum_by(code)_(task:errors:rate10s{job="s...-8      12.000 ± 0%   3.000 ± 0%  -75.00% (p=0.002 n=6)
ExprString/max(_100_*_(1_-_avg_by(instance)_(irate(...-8      34.000 ± 0%   5.000 ± 0%  -85.29% (p=0.002 n=6)
ExprString/http_requests_total{job="api-server",_gr...-8      30.000 ± 0%   8.000 ± 0%  -73.33% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...-8       58.00 ± 0%   19.00 ± 0%  -67.24% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...#01-8   58.000 ± 0%   9.000 ± 0%  -84.48% (p=0.002 n=6)
ExprString/label_replace(testmetric,_"dst",_"destin...-8      22.000 ± 0%   4.000 ± 0%  -81.82% (p=0.002 n=6)
geomean                                                        31.11        6.592       -78.81%
```