@charleskorn 2025-06-10
>I'm not convinced we need maximum complexity; most of the benefit is achieved with the first two changes.

First two or three seem reasonable to me, I'm not sure the extra complexity of the final change is worth it, and I worry that we could unintentionally lose the benefit of it if something somewhere changes the breaks the ordering of the matchers.